Moved ComponentRange class into the StructureManager class. Moved StructureManager-related convenience functions from StructureComponent.h into StructureManager.h.
